requirements.txt
tensorflow-gpu==2.0.0-alpha0
matplotlib==3.0.3
numpy==1.16.2
pandas==0.24.2
scikit-learn==0.20.3